requirements.txt
streamlit
numpy
pickle-mixin
pillow
scikit-learn